restore LogLvl after ComfyUI import
