+++ visionatrix/etc.py
@@ -1,0 +1,24 @@
+import logging
+
+
+def get_log_level(log_level_str):
+    """Convert log level string to logging module log level."""
+    log_levels = {
+        "DEBUG": logging.DEBUG,
+        "INFO": logging.INFO,
+        "WARNING": logging.WARNING,
+        "ERROR": logging.ERROR,
+        "CRITICAL": logging.CRITICAL,
+    }
+    return log_levels.get(log_level_str.upper(), logging.INFO)
+
+
+def get_higher_log_level(current_level):
+    level_mapping = {
+        logging.DEBUG: logging.INFO,
+        logging.INFO: logging.WARNING,
+        logging.WARNING: logging.ERROR,
+        logging.ERROR: logging.ERROR,
+        logging.CRITICAL: logging.CRITICAL,
+    }
+    return level_mapping.get(current_level, logging.WARNING)
